requirements.txt
flask
flask-cors
python-dotenv
supabase
google-generativeai
requests
gunicorn
